update forgot password screen structure
